# api.yaml
swagger: "2.0"
info:
  description: "This is the auth service for Muxi-X(Called 木犀内外)'s API document, please make sure you're using the python version server service."
  version: "2.0.0"
  title: "Auth Service v2"
  termsOfService: "http://muxixyz.com/api/auth"
  contact:
    email: "muxistudio@qq.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "pass.muxi-tech.xyz"
basePath: "/auth/api"
...
schemes:
  - "https"
  - "http"
paths:
...
  /admin/members:
    get:
      tags:
        - "admin"
      summary: "list muxi members"
      description: "需要本地管理员账号的 access token。"
      parameters:
        - in: header
          name: token
          description: "OAuth access token for a local admin user."
          required: true
          type: string
        - in: query
          name: page
          required: false
          type: integer
          description: "从 1 开始，默认 1"
        - in: query
          name: page_size
          required: false
          type: integer
          description: "默认 20，上限 100，超出会被压到 100"
        - in: query
          name: group
          required: false
          type: string
          description: "按组别精确过滤"
        - in: query
          name: query
          required: false
          type: string
          description: "模糊匹配 real_name / student_id / username / email"
      responses:
        200:
          description: ok
          schema:
            $ref: "#/definitions/MemberList"
        401:
          description: "Invalid or missing access token."
        403:
          description: "Admin permission is required."
    post:
      tags:
        - "admin"
      summary: "grant muxi membership to an existing account"
      description: "把已有账号授权为木犀成员。学号唯一性在应用层校验，空学号豁免。"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: header
          name: token
          description: "OAuth access token for a local admin user."
          required: true
          type: string
        - in: body
          name: data
          schema:
            required:
              - user_id
              - group
            properties:
              user_id:
                type: integer
              real_name:
                type: string
              student_id:
                type: string
              group:
                type: string
                description: "Frontend / Backend / Design / Product / Operation 之一"
              join_year:
                type: integer
              personal_blog:
                type: string
              github:
                type: string
              zhihu:
                type: string
      responses:
        200:
          description: ok
          schema:
            $ref: "#/definitions/MemberProfile"
        400:
          description: "组别不在白名单（20603）、学号已被占用（20604）、或者该账号已经是成员（20602）。"
        401:
          description: "Invalid or missing access token."
        403:
          description: "Admin permission is required."
        404:
          description: "user_id 对应的账号不存在（20102）。"
  /admin/members/{user_id}:
    put:
      tags:
        - "admin"
      summary: "update a muxi member profile"
      description: "请求体里所有字段可选，零值表示不改这一项。因为空值等于不修改，所以这个接口没法把字段改回空串。"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: header
          name: token
          description: "OAuth access token for a local admin user."
          required: true
          type: string
        - in: path
          name: user_id
          required: true
          type: integer
        - in: body
          name: data
          schema:
            properties:
              real_name:
                type: string
              student_id:
                type: string
              group:
                type: string
              join_year:
                type: integer
              personal_blog:
                type: string
              github:
                type: string
              zhihu:
                type: string
      responses:
        200:
          description: ok
          schema:
            $ref: "#/definitions/MemberProfile"
        400:
          description: "组别不在白名单（20603）或学号已被别的成员占用（20604）。"
        401:
          description: "Invalid or missing access token."
        403:
          description: "Admin permission is required."
        404:
          description: "该账号不是木犀成员（20601）。"
    delete:
      tags:
        - "admin"
      summary: "revoke a muxi membership"
      description: "删除成员档案，并撤销该用户已签发的 OAuth token。注意撤销只能让 refresh token 立即失效：已签发的 access token 是自包含的 JWT，在剩余有效期内（最长 30 天）仍可通过本地验签使用，要做到即时失效，下游必须回查 GET /user。"
      parameters:
        - in: header
          name: token
          description: "OAuth access token for a local admin user."
          required: true
          type: string
        - in: path
          name: user_id
          required: true
          type: integer
      responses:
        200:
          description: ok
        401:
          description: "Invalid or missing access token."
        403:
          description: "Admin permission is required."
        404:
          description: "该账号不是木犀成员（20601）。"
...
definitions:
...
  MemberProfile:
    type: "object"
    properties:
      user_id:
        type: integer
      real_name:
        type: string
      student_id:
        type: string
      group:
        type: string
        description: "管理端写入时只接受 Frontend / Backend / Design / Product / Operation；读出来可能是「前端」「Android」这类历史自由文本，不要拿白名单过滤。"
      join_year:
        type: integer
        description: "老数据里 timejoin 解析不出年份时为 null。"
      personal_blog:
        type: string
      github:
        type: string
      zhihu:
        type: string
      created_at:
        type: string
        description: "RFC3339。历史成员取的是迁移时刻，不是真实入队时间。"
      updated_at:
        type: string
  MemberListItem:
    allOf:
      - $ref: "#/definitions/MemberProfile"
      - type: "object"
        properties:
          username:
            type: string
          email:
            type: string
          avatar_url:
            type: string
  MemberList:
    type: "object"
    properties:
      list:
        type: array
        items:
          $ref: "#/definitions/MemberListItem"
      total:
        type: integer
      page:
        type: integer
      page_size:
        type: integer